refactor(core): further decouple LEVM and rEVM implementations.
