+++ crates/vm/backends/revm/mod.rs
@@ -1,16 +1,21 @@
-use super::constants::{
+pub mod aux;
+pub(crate) mod db;
+pub mod execution_db;
+pub mod execution_result;
+
+use super::BlockExecutionResult;
+use crate::constants::{
     BEACON_ROOTS_ADDRESS, CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS, HISTORY_STORAGE_ADDRESS,
     SYSTEM_ADDRESS, WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
 };
-use super::BlockExecutionResult;
 use crate::spec_id;
 use crate::EvmError;
-use crate::EvmState;
-use crate::ExecutionResult;
+use db::EvmState;
 use ethrex_common::types::AccountInfo;
 use ethrex_common::{BigEndianHash, H256, U256};
 use ethrex_storage::{error::StoreError, AccountUpdate};
 
+use execution_result::ExecutionResult;
 use revm::db::states::bundle_state::BundleRetention;
 use revm::db::AccountStatus;
 use revm::{
