+++ crates/vm/constants.rs
